working on authorizing models and transaction controller

+++ app/Http/Controllers/Core/BudgetController.php
@@ -7,10 +7,13 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\Core\CreateBudgetRequest;
 use App\Http\Requests\Core\UpdateBudgetRequest;
 use App\Http\Resources\Core\BudgetResource;
+use App\Models\Budget;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\Request;
 
 class BudgetController extends Controller
 {
+    use AuthorizesRequests;
 
     public function __construct(protected  BudgetServiceInterface $budgetService)
     {
@@ -31,23 +34,27 @@ class BudgetController extends Controller
 
     public function show(int $id)
     {
+        $this->authorize('view', Budget::class);
         $result = $this->budgetService->findOrFail($id);
         return new BudgetResource($result);
     }
 
     public function create(CreateBudgetRequest $request)
     {
+        $this->authorize('create', Budget::class);
         $result = $this->budgetService->create($request->validated());
         return new BudgetResource($result);
     }
 
     public function update(int $id,UpdateBudgetRequest $request)
     {
+        $this->authorize('update', Budget::class);
         $result = $this->budgetService->update($id, $request->validated());
         return new BudgetResource($result);
     }
 
     public function delete(int $id){
+        $this->authorize('delete', Budget::class);
         $deleted = $this->budgetService->delete($id);
         if($deleted){
             return response()->json(["message"=>"Category deleted successfully"]);
